refactor: improve clarity and type-safety in various calculations and loops across multiple classes

+++ src/main/java/com/peco2282/bcreborn/common/builder/BuildingItem.java
@@ -152,7 +152,7 @@ public class BuildingItem implements IBufferSerializable, IBuildingItem {
     float displayPortion = timeSpan / tickDuration;
 
     if (lifetimeDisplay - lifetime <= 1.0) {
-      lifetimeDisplay += 1.0 * displayPortion;
+      lifetimeDisplay += (float) (1.0 * displayPortion);
     }
   }
